+++ edgelet/support-bundle/src/lib.rs
@@ -6,7 +6,7 @@
 
 mod error;
 mod logs;
-// mod support_bundle;
+mod support_bundle;
 
 pub use crate::error::{Error, ErrorKind};
 pub use crate::logs::pull_logs;
